Remove specific Jedis implementation from RedisCache
